extracted collector class
